refactor(daemon): name the credential port into the agent-runtime family
The port exists only to store and resolve credentials for runtime instances:
its sole importer inside the daemon is agent-runtime-instances.ts, and the GUI
broker uses it for the same runtime instances. It belongs to the same naming
family as agent-runtime-instances.ts and agent-runtime-stream.ts, and sat
outside it only because the file was new.
